refactor(mcp-actions): split plugin bootstrap into facade modules
Keep mcp plugin startup focused on orchestration by delegating MCP route wiring and OIDC discovery route registration to dedicated modules with unchanged behavior.
